PDF-Search-Assistant/requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
torch>=2.0.0
transformers>=4.40.0
